% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.piecewise.R
\name{mu.piecewise}
\alias{mu.piecewise}
\title{Produce mean matrix for CPL}
\usage{
mu.piecewise(doses)
}
\arguments{
\item{doses}{a numeric vector of specified doses.}

\item{k}{a numeric value specifies the number of active doses.}
}
\value{
mean matrix for CPL
}
\description{
This function is a help function to automatrically produce mean matrix for CPL.
}
\examples{
k<-4
doses<-seq(0,k)
mu.piecewise(doses)

}

