% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optC.R
\name{optC}
\alias{optC}
\title{Produce optimal contrast}
\usage{
optC(mu, Sin)
}
\arguments{
\item{mu}{numeric values}

\item{Sin}{inverse of covariance matrix}
}
\value{
optimal contrast
}
\description{
This function is a help function to automatrically produce mean vector for linear.
}

