% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{trendpow}
\alias{trendpow}
\title{Calculate power for multiple contrast tests}
\usage{
trendpow(trueMu = NULL, trueModel = NULL, n = NULL, sigma = NULL,
  S = NULL, df = NULL, evalDoses = NULL, trend = c("CPL", "MCP Mod",
  "Aug Williams", "Williams", "Dunnett", "Linear"), evalModel = NULL,
  alternative = "one.sided", alpha = 0.05)
}
\arguments{
\item{trueMu}{mean values in all doses groups including placebo. It could be a vector of arbitrary values or an object from CPL.}

\item{trueModel}{an object of class "Mods", see Mods in DoseFinding package for detail. It is used to define mean values. Specify trueMu if you want to an arbitrary mu or an object from CPL,
or specify trueModel if you want to use an object from MCP Mods. If trueMu and trueModel both are missing, then trueMu will be an object from MCP Mod if trend is specified as
MCP Mod or an object from CPL for all other trends.}

\item{n}{a numeric vector of sample sizes in all doses groups including placebo.}

\item{sigma}{a single numeric value of residual standard deviation. It is assumed computation are made for a normal homoscedastic model with group sample sizes given by n and residual standard deviation sigma,
i.e. the covariance matrix used for estiamtes is thus sigma^2*diag(1/n)}

\item{S}{convariance matrix for the estimates, either n and sigma or S need to be specified.}

\item{df}{specify degrees of freedom for multivariate t. If this argument is missing, df=0 corresponding multivariate normal.}

\item{evalDoses}{a numeric vector of specified doses. When this argument is missing, doses are 0 to k by 1. You can assign less levels of doses. True mu will be adjusted correspondingly.}

\item{trend}{a single character string, determining the trend for the multiple contrast trend test, one of "CPL", "MCP Mod","Aug Williams", "Williams", "Dunnett", "Linear".
Note that when "MCP_Mod" is specified for trend, model needs to be specified.}

\item{evalModel}{an object of calss "Mods" if you choose trend to be MCP_Mod trend test, see Mods for details.}

\item{alternative}{a single character string, specifying the alternative for the multiple contrast trend test. If this argument is missing, alternative="one.sided".}

\item{alpha}{significance level to use. If this argument is missing, alpha=0.05.}
}
\value{
true mean, optimal contrast, mean matrix, power.
}
\description{
Calculate power for a multiple contrast test for a specified trend type.
}
\examples{
N<-150
n0<-30
n1<-30
k<-5
n<-c(n0, rep(n1,k))
#doses
ds<-seq(0,5)


models<-Mods(emax=c(log(1+2*0.01),log(1+2*0.25),log(1+2*4)),
          linear=NULL,
          exponential=c(log(1+2*0.25), log(1+2*0.75), log(1+2*3.5)),
          sigEmax=rbind(c(log(1+2*1),25), c(log(1+2*5),25), c(log(1+2*15),25),c(log(1+2*1),5), c(log(1+2*5),5), c(log(1+2*15),5) ),
          doses=ds, placEff=0, maxEff=0.7)


MCP_models<-Mods(emax=c(log(1+2*0.01),log(1+2*0.25),log(1+2*4)),
                linear=NULL,
                exponential=c(log(1+2*0.25), log(1+2*0.75), log(1+2*3.5)),
                sigEmax=rbind( c(log(1+2*1),5), c(log(1+2*5),5), c(log(1+2*15),5) ),
                doses=ds, placEff=0, maxEff=0.7)


trendpow<-trendpow(trueModel=models,n=rep(30,6), sigma=1,evalDoses=ds, alpha=0.05, trend="MCP Mod", evalModel=MCP_models)

mufrmCPL<-mu.piecewise(ds)

trendpow<-trendpow(trueMu=mufrmCPL,n=rep(30,6), sigma=1,evalDoses=ds, alpha=0.05, trend="MCP Mod", evalModel=MCP_models)
}

